[Improve][Core][Starter] Refactor seatunnel starter
