+++ seatunnel-core/seatunnel-core-starter/src/main/java/org/apache/seatunnel/core/starter/utils/CommandLineUtils.java
@@ -19,7 +19,7 @@ package org.apache.seatunnel.core.starter.utils;
 
 import static org.apache.seatunnel.core.starter.constants.SeaTunnelStarterConstants.USAGE_EXIT_CODE;
 
-import org.apache.seatunnel.core.starter.command.AbstractCommandArgs;
+import org.apache.seatunnel.core.starter.command.CommandArgs;
 import org.apache.seatunnel.core.starter.command.UsageFormatter;
 
 import com.beust.jcommander.JCommander;
@@ -31,11 +31,11 @@ public class CommandLineUtils {
         throw new UnsupportedOperationException("CommandLineUtils is a utility class and cannot be instantiated");
     }
 
-    public static <T extends AbstractCommandArgs> T parse(String[] args, T obj) {
+    public static <T extends CommandArgs> T parse(String[] args, T obj) {
         return parse(args, obj, null, false);
     }
 
-    public static <T extends AbstractCommandArgs> T parse(String[] args,
+    public static <T extends CommandArgs> T parse(String[] args,
                                                           T obj, String programName,
                                                           boolean acceptUnknownOptions) {
         JCommander jCommander = JCommander.newBuilder()
