+++ seatunnel-core/seatunnel-starter/src/main/java/org/apache/seatunnel/core/starter/seatunnel/args/ServerCommandArgs.java
@@ -21,31 +21,10 @@ import org.apache.seatunnel.core.starter.command.CommandArgs;
 
 import com.beust.jcommander.Parameter;
 
-import java.util.List;
-
-public class ServerCommandArgs implements CommandArgs {
-
-    /**
-     * Undefined parameters parsed will be stored here as seatunnel engine command parameters.
-     */
-    private List<String> seatunnelParams;
-
+public class ServerCommandArgs extends CommandArgs {
     @Parameter(names = {"-cn", "--cluster"},
         description = "The name of cluster")
     private String clusterName = "seatunnel_default_cluster";
-
-    @Parameter(names = {"-h", "--help"},
-        help = true,
-        description = "Show the usage message")
-    private boolean help = false;
-
-    public boolean isHelp() {
-        return help;
-    }
-
-    public void setHelp(boolean help) {
-        this.help = help;
-    }
 
     public String getClusterName() {
         return clusterName;
@@ -53,13 +32,5 @@ public class ServerCommandArgs implements CommandArgs {
 
     public void setClusterName(String clusterName) {
         this.clusterName = clusterName;
-    }
-
-    public List<String> getSeatunnelParams() {
-        return seatunnelParams;
-    }
-
-    public void setSeatunnelParams(List<String> seatunnelParams) {
-        this.seatunnelParams = seatunnelParams;
     }
 }
